rename file and function
